add custom valdiator

+++ src/TradingApp.TradingAdapter/Models/CombinedQuote.cs
@@ -1,3 +1,3 @@
 ﻿namespace TradingApp.TradingAdapter.Models;
 
-public record CombinedQuote(Quote Ohlc, double? Rsi, double? Sma);
+public record CombinedQuote(Quote Ohlc, decimal? Rsi, decimal? Sma);
