+++ tests/TradingApp.Application.Test/Quotes/GetStooqCombinedQuotesCommandHandlerTests.cs
@@ -41,7 +41,7 @@ public class GetStooqCombinedQuotesCommandHandlerTests
     {
         //Arrange
         StooqProvider.Setup(_ => _.GetQuotes(new GetQuotesRequest(command.TimeFrame, command.Asset, null))).ReturnsAsync(Result.Ok(quotes));
-        var values = Enumerable.Range(0, quotes.Count()).Select(_ => (double?)new Random().NextDouble()).ToList();
+        var values = Enumerable.Range(0, quotes.Count()).Select(_ => (decimal?)new Random().NextDouble()).ToList();
         Evaluator.Setup(_ => _.GetRSI(It.IsAny<IEnumerable<Quote>>(), It.IsAny<int>())).Returns(values);
         //Act
         var result = await _sut.Handle(command, CancellationToken.None);
